pyproject.toml
[project]
name = "todolist-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]
